Completing menus, send email etc
* Finishing off implementation of Options and Context menus in all
Activities
* Implemented "Share Shopping List" via email, sms and other
applications that support the Intent.ACTION_SEND.
* Group and item list separators reverted back to using the same.
Android seems to use the group indicator only if it is an expanded
group. This does not help what I was trying to achieve namely a look
where groups would look like plateaus.
* Fixed a bug in edit item where the item would not get renamed if the
new name was not already present in the catalog table. This bug was
introduced because of a previous refactoring.
* Fixed the addLeftAlignedButton in the CustomTitlebarWrapper. Was an
error on my part.
*  Implemented Showing/Hiding of Done items.
* Added a preference for "Show pending Items Count in Shoping List".