Fetching The Pending Total from Backend
As the entirety of a Shopping List may not have been fetched in one
shot, it is impossible to calculate the Total Pending Cost in the front
end. Using the backend for such a calculation. An optimization made is
that when items are added/removed/edited/marked done, the total is
updated in the frontend without a backend query.

Also removing the display of Total Amount Spent as it will only keep
increasing and does not have much value as a statistic. The details
will anyway be available in the Reports.