Introducing bar code scanning!
We're using zxing application with intents for scanning the barcode. For now, item details are fetched for the scanned item using it's UTC and google shopping api's. The back end needs to prepare to store barcodes as well.

Major, revamp of the AsyncInvokeUTLTask API's inorder to support both http/https and get/post requests. Things would have been simpler if min_sdk was 8 and above.