Searching for items based on barcode moved to backend
The implementation is now two fold. First the barcode will be checked
in the database. If it is found, details about that particular row will
be returned. If it is not found, google will be searched. Currently
there a 25K limit on number of queries allowed per day per free account.