We're breaking new grounds in usability here. Users can now swipe right on an item and the item will be marked done. The other option is to click on an item, wait for the quick action popup to show up, then click done.
We using a regular GestureDetector along with SimpleGestureListener for detecting fling across an item. There's a problem with the onChildClick listener in that it is always called whether we detect a fling on not leading to both a fling and item click being fired. This obviously, is not desirable. Hence we do not use the ListView.onClickListener. Instead we've simulated onChildClick thorough the SimpleGestureListener and it seems to work fine.

Yippie!

TBD: Finalize of the behaviorial aspects and remove both "Hide When Done" and "Move done Items to bottom of list" preferences. We could go for a tristate implementation where items can be pending/purchased/archived state. We always display the first two type of item and allow a "shake" to determine when purchased items should be archived.