Checking in code for detecting shakes. The plan is to have users "shake" the device to hide all done items. The code doesn't seem to be working but checking in anyway as I expect to get it to work as soon as I can spend some time on it.
Removing the "Shuffle_Done_Items" preference as it may cause confusion with the "Delete_Bought_Items" preference. The default behavior is to move done items to the bottom of the parent category. It would have been nice if we could animate this shuffle. However, it seems to be a lot of work.