Managed to fix the problem with the tips dialog: http://stackoverflow.com/questions/11189401/styled-background-used-in-layout-of-transparent-popupwindow-not-working/11274257#11274257
I ultimately narrowed down the problem to my passing the context returned by getApplicationContext() to the PopupWindow constructor. If I pass LoginActivity.this as the context, things work as expected and I'm able to change backgrounds according to theme selected in the parent activity.

Still beats me why getApplicationContext() would not have the theme I set on it. Food for thought.